Clean up dead code surfaced by fallow
Dead code removed:
- Delete src/components/deploy/PromoteDialog.tsx (superseded by PromoteTab).
- Remove unused resolveCommit, listPromotionOptions from api/endpoints.ts
  and the now-orphaned PromotionOption type.
- Remove unused ciStatusDotClass from lib/status-colors.ts.
- Drop unused `export` keyword from AnchorKind (AnchorEdgesCard.tsx) and
  DayBucketKey (opsLogHelpers.ts).
- Remove unused type exports: ApiStatus, AdminUserUpdate, AuthState,
  DashboardStats, ProjectRelationshipSummary, ProjectService,
  ProjectServiceCreate, ServiceAccountUpdate,
  ServiceApplicationSecretsUpdate, View, ViewChangeConfig (types/index.ts),
  SendMessageRequest, SSEClientActionEvent, SSEDoneEvent, SSEErrorEvent,
  SSETextEvent, SSEToolInputEvent, SSEToolUseStartEvent (types/assistant.ts).

Architecture refactors to break import cycles:
- Extract queryClient from main.tsx into src/lib/queryClient.ts; update
  CommandBar.tsx accordingly. Kills the 8 cycles through main.tsx.
- Extract getTypeSwatch from BlueprintManagement.tsx into a new
  blueprints/typeSwatch.ts; update BlueprintDetail.tsx to import from
  there. Kills the BlueprintManagement <-> BlueprintDetail cycle.

False-positive suppressions:
- Add TeamScoreSeries to the existing endpoints.ts ignoreExports list
  (same class as the other reports/ types fallow can't see transitively).
- Add an overrides entry turning off duplicate-exports for
  src/lib/icon-sets/*.ts (each module intentionally exports `iconSet`).
- Inline fallow-ignore comment for the vi.mock unresolved-import false
  positive in DeploymentRunWatcher.test.tsx.
